Update the Gradle build tools to 23.0.2 (from 23.0.1)
When running "react-native run-android", we could see the following suggestion: "Running dex in-process requires build tools 23.0.2. For faster builds update this project to use the latest build tools." Therefore, we upgraded the build tools to 23.0.2 in order to make the builds faster. I also upgraded the heap size to 1536M (from the default 1024M) since the Gradle Daemon cannot run dex in-process with less than 1536M.