Add more steps for running the example project to fix an issue with npm >= 5
npm 5 changed the way `npm install` installs and links local modules. Previously, if a dependency in package.json was using the `file: <relative/path/to/file` protocol, npm install would copy the folder inside node_modules as if it were downloading a module from the npm registry. But since npm 5, `npm install` only creates a symlink to the module. Since the react-native packager doesn't work well with symlinks, it cannot find the react module when running the app. A workaround is to install "instal-local", which is a tool made specifically to fix this npm 5 problem. More info on the issue: https://stackoverflow.com/questions/44624636/npm-5-install-folder-without-using-symlink