Fix the Android model translation
jPCT doesn't really work like most 3D engines, so we have to do a lot of
extra manipulations. For example, all the operations (scale, rotation,
translation) are done in local space. This means that when points are
really far from the origin in a model file, scaling it down won't
proportionally bring those points closer to the origin. Instead, the
points will scale down on their relative center, which is not what we
wanted.