Add tint color for Android
To tint the texture on Android, we use the jPCT ITextureEffect
interface. By implementing this interface, we have access to the raw
bitmaps of the texture and can multiply their value by the tint color.
If there's no texture, we simply create a new 1x1 texture with the tint
color.