Fix a tint bug on iOS
When the user didn't explicitly specify a component of the tint color on
the javascript side (e.g. green), the component would have a value of 0
by default. Since we multiply the tint with the texture color, the tint
would kill the color component even though the user didn't specify it.
By setting the default as 1.0 instead, nothing happens if the user
voluntarily omits the component.